Initial commit
First attempt at training a neural network to convert an image to ascii.
This is meant as a launching point for more work that makes the thing
actually functional

At this point we have something that can train a network that attempts
to output a character that somewhat represents the brightness value of a
rectangular area of pixels. This is not producing meaningful results in
it's current state

There is also a way to visualize what the expected output would be by
printing to the screen. This does not currently match anything we've
trained

The initial network architecture is an attempt at the dumbest thing that
could possibly work. If we train based off brightness of a sample, in
theory no hidden layers should be needed as the correct character should
be able to be determined by just checking how bright the sample is.

Code is rough, and that's ok. It's better to get something committed so
we can produce meaningful diffs going forward