Add label visualization UI
As we tune our labeling metrics, it will be important to see how
different affect the output. Add a simple web ui that shows the input
and textified images, as well as the scores for each glyph

Web is used for the visualization because it's effectively zero
dependency. We have an http server, so we might as well use it

Visualizations can be run by starting visualize.py and going to
localhost:9999/index.html

Bonus fix: Remove 0 width glyphs from