Update labeling metric
The existing labeling metric did not produce very good results. One of
the goals of this project is to accomplish sub-glyph accuracy on the
textified image.

Previously we were just using overall brightness of a sample from the
source image vs brightness of the glyph. This does not take into account
positioning of the glyph within the sample at all.

After experimentation, blurring the glyph a little, then directly
comparing the blurred glyph with the sample, and then weighting the
results a little bit with overall brightness seems to give pretty decent
results

Implement this as our default metric

The rendered glyph size had to be updated a little. The blur has a
minimum area of effect of 1 pixel. If we want a smaller impact, we need
more pixels. Arbitrarily bumped to a number that felt like it produced
good results on our single sample image

See https://www.youtube.com/watch?v=2t-egoVAfWU for some of the
alternative approaches that were tried and discarded