Add visualization support for multiple labeling metrics
As we work towards making something actually useful, we need to create
labels for our images that are half way decent. The brightness metric
was a placeholder and does not look very good.

Add support for multiple scoring functions, and show them in the
visualizer UI

Add a sample "diff" metric to prove the concept

* HTML dropdown added for selecting the labeling metric
* JS now requests the supported metrics from python server
* For each sample related API, a metric query param has been added
* On launch, the visualizer computes the output image for each labeling
  metric, and serves the one requested with the metric query param