Train/visualize multiple images
Replace existing ImgSampler with a NestedDirImageLoader. This will glob
search the input directory for all images and use those for training

Visualization script needed to be updated to support visualizing more
than one image. We no longer cache output images and compute outputs on
the fly. This results in increased cost for each image, but it's
manageable so we leave it for now.

Sample selection for multiple images got a little complex, width/height
ratios may change after css rules are applied. On click instead of using
the width/height ratio seen on image load, use the latest and greatest