Separate glyph rendering from label generation
Refactoring to make future code easier to work with. Specifically we do
not want the ImgSampler to be in charge of glyph cache generation. We
might want to generate a glyph cache from different inputs, or configure
font selection, etc.

Shuffle glyph cache generation code around to be external to ImgSampler,
and clean some stuff up while we're in the area

* Previously there was double caching happening, once with the glyph
  after it came out of freetype, once in torch stamps. We don't have to
  do both, remove the freetype layer
* Now that the glyph cache is external, only store the blurred glyph
  cache in the img sampler. The original is unused
* Split out a text -> image generator that uses the original glyph
  cache, deduplicating image rendering code that lived in both the
  visualizer and the training script (and soon to be other scripts as
  well)