Preprocess image to improve definition
Glyph brightness distribution is skewed to be pretty dark, which means
that there is very little glyph (and therefore edge) variety for most of
the color range. If we adjust our color range to better match the glyph
distribution we get better definition of edges in the scene

While we're sacrificing brightness for edge detail, apply a local
contrast enhancement from imagemagick. This way we get more local
definition of edges without heavily sacrificing global brightness.